requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
chromadb==1.0.15
openai==1.99.9
google-generativeai>=0.7.0
sentence-transformers>=3.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
jinja2==3.1.6
beautifulsoup4==4.13.4
markdownify==1.1.0
pypdf==5.8.0
pytest==8.4.1
httpx==0.28.1
